+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 2020-04-10 v2.2.0-SNAPSHOT
+* [MODSOURMAN-276](https://issues.folio.org/browse/MODSOURMAN-276) Created endpoint for retrieving MARC record by instance id
 
 ## 2020-04-23 v2.1.3
 * [MODSOURMAN-303](https://issues.folio.org/browse/MODSOURMAN-303) Add actual state on creating record.